Start - 1st commit
I created all the files that may be used for this project. I went ahead and made the html boilerplate, linked the style page and JS to handle functionality, both are working. I created a CSV to hopefully be able to log the job number, operator 2-letter initials and start time and date. I created a basic form in the html to hopefully be able to take those input values and append them to the CSV as independent records.