Update
Refactoring the massive function that I had at the start. Now each input will have it's own function that passes an argument or value to the output function that at first will just print out to console. Eventually I will set it up to save to either a text or CSV (ideally).