Update
I created a skeleton of what I would like the steps to look like...added a ton to the JS for functionality, will need to look up a better way to hide divs that doesn't require a dedicated function to do so. I'll have to add in the steps and optional selections. Also Images when I get time.